requirements.txt
opencv-python
face-recognition
facenet-pytorch
numpy
pandas
torchvision
Pillow